code cpleanup
