+++ tasks_pkg/__init__.py
@@ -3,3 +3,5 @@ from . import helpers
 from . import metal
 from . import network
 from . import cluster
+from . import k8s_context
+from . import equinix_metal
