+++ templates/default.yaml
@@ -80,7 +80,7 @@ spec:
               ip: ${TOEM_CP_ENDPOINT}
               equinixMetal:
                 apiToken: ${PACKET_API_KEY}
-        - op: add
+        - op: add # https://www.talos.dev/v1.4/kubernetes-guides/network/kubespan/#enabling-for-an-existing-cluster
           path: /machine/network/kubespan
           value:
             enabled: true
